BIG UPDATE
Một số cơ chế đã ổn định như load map, load quái, xử lí va chạm (tương đối)
đang làm chuyển room
bấy lâu nay tui copy sai folder git mà tưởng nó tự động update :|

+++ Metroid_Game/WinMain/Item.cpp
@@ -33,7 +33,7 @@ void Item::InitSprites(LPDIRECT3DDEVICE9 d3ddv, LPDIRECT3DTEXTURE9 texture)
 void Item::Init(float posX, float posY)
 {}
 
-void Item::Update(float t)
+void Item::Update(int t)
 {
 	if (!isActive)
 		return;
